requirements.txt.txt
pillow
opencv-python
numpy
pyqt5